[ISSUE-44] Modularized user info form
